+++ test/spec/cider/nrepl/middleware/info_spec_test.clj
@@ -23,7 +23,7 @@
 
 (deftest integration-test
   (testing "spec info on a normal function with spec"
-    (let [response (session/message {:op "info" :symbol "ranged-rand" :ns "cider.nrepl.middleware.info-spec-test"})]
+    (let [response (session/message {:op "info" :sym "ranged-rand" :ns "cider.nrepl.middleware.info-spec-test"})]
       (is (= (:status response) #{"done"}))
       (is (= (:ns response) "cider.nrepl.middleware.info-spec-test"))
       (is (= (:name response) "ranged-rand"))
@@ -40,7 +40,7 @@
                                       ["clojure.core/fn" ["%"] ["clojure.core/<" [":ret" "%"] ["clojure.core/->" "%" ":args" ":end"]]]]]))))
   (testing "same name testing function without a spec"
       ;; spec is not defined for this function
-    (let [response (session/message {:op "info" :symbol "same-name-testing-function" :ns "cider.test-ns.first-test-ns"})]
+    (let [response (session/message {:op "info" :sym "same-name-testing-function" :ns "cider.test-ns.first-test-ns"})]
       (is (= (:status response) #{"done"}))
       (is (= (:ns response) "cider.test-ns.first-test-ns"))
       (is (= (:name response) "same-name-testing-function"))
@@ -50,7 +50,7 @@
       (is (nil? (:spec response)))))
 
   (testing "spec info on clojure.core/let"
-    (let [response (session/message {:op "info" :symbol "let" :ns "cider.nrepl.middleware.info-spec-test"})]
+    (let [response (session/message {:op "info" :sym "let" :ns "cider.nrepl.middleware.info-spec-test"})]
       (is (= (:status response) #{"done"}))
       (is (= (:ns response) "clojure.core"))
       (is (= (:name response) "let"))